Summary table working again
