+++ data/results-history.json
@@ -1,6 +1,6 @@
 [
   {
-    "date": "2026-06-12T14:50:38.083Z",
+    "date": "2026-06-14T12:51:24.099Z",
     "testCount": 412,
     "minifiers": {
       "clean-css": {
