+++ lib/helpers.js
@@ -31,3 +31,7 @@ export const formatMs = function (duration) {
   const time = prettyMilliseconds(duration, options);
   return time;
 };
+
+export const html = function (strings, ...values) {
+  return String.raw({ raw: strings }, ...values);
+};
